# perf: Switch SafeEmitter to use a map