--- a/lib/linter/safe-emitter.js
+++ b/lib/linter/safe-emitter.js
@@ -30,23 +30,23 @@
  * @returns {SafeEmitter} An emitter
  */
 module.exports = () => {
-	const listeners = Object.create(null);
+	const listeners = new Map();
 
 	return Object.freeze({
 		on(eventName, listener) {
-			if (eventName in listeners) {
-				listeners[eventName].push(listener);
+			if (listeners.has(eventName)) {
+				listeners.get(eventName).push(listener);
 			} else {
-				listeners[eventName] = [listener];
+				listeners.set(eventName, [listener]);
 			}
 		},
 		emit(eventName, ...args) {
-			if (eventName in listeners) {
-				listeners[eventName].forEach(listener => listener(...args));
+			if (listeners.has(eventName)) {
+				listeners.get(eventName).forEach(listener => listener(...args));
 			}
 		},
 		eventNames() {
-			return Object.keys(listeners);
+			return Array.from(listeners.keys());
 		},
 	});
 };