@mdjermanovic 2025-04-29
I'm getting the following on Windows, Node.js v22.14.0.

Before:

```
$ npm run test:performance

> eslint@9.25.1 test:performance
> node Makefile.js perf


Loading:
  Load performance Run #1:  149.3193ms
  Load performance Run #2:  144.6692ms
  Load performance Run #3:  145.4597ms
  Load performance Run #4:  146.2858ms
  Load performance Run #5:  148.305ms

  Load Performance median:  146.2858ms


Single File:
  CPU Speed is 2803 with multiplier 13000000
  Performance Run #1:  2176.8739ms
  Performance Run #2:  1943.0212999999999ms
  Performance Run #3:  1999.7469ms
  Performance Run #4:  1991.4533999999999ms
  Performance Run #5:  1992.2753ms

  Performance budget ok:  1992.2753ms (limit: 4637.887977167321ms)


Multi Files (450 files):
  CPU Speed is 2803 with multiplier 39000000
  Performance Run #1:  5097.2704ms
  Performance Run #2:  5142.8913ms
  Performance Run #3:  5108.4245ms
  Performance Run #4:  5204.0834ms
  Performance Run #5:  5114.362ms

  Performance budget ok:  5114.362ms (limit: 13913.663931501962ms)
```

After:

```
$ npm run test:performance

> eslint@9.25.1 test:performance
> node Makefile.js perf


Loading:
  Load performance Run #1:  149.9871ms
  Load performance Run #2:  143.9361ms
  Load performance Run #3:  141.0293ms
  Load performance Run #4:  142.9743ms
  Load performance Run #5:  139.0067ms

  Load Performance median:  142.9743ms


Single File:
  CPU Speed is 2803 with multiplier 13000000
  Performance Run #1:  2118.4715ms
  Performance Run #2:  1983.3435ms
  Performance Run #3:  1980.6439ms
  Performance Run #4:  1955.1788000000001ms
  Performance Run #5:  1946.2447ms

  Performance budget ok:  1980.6439ms (limit: 4637.887977167321ms)


Multi Files (450 files):
  CPU Speed is 2803 with multiplier 39000000
  Performance Run #1:  5112.8065ms
  Performance Run #2:  5079.0783ms
  Performance Run #3:  5129.6498ms
  Performance Run #4:  5202.0729ms
  Performance Run #5:  5219.4505ms

  Performance budget ok:  5129.6498ms (limit: 13913.663931501962ms)
```