@nzakas 2025-04-29
Your computer is faster than mine. :smile: Surprised to not see any difference. I'm on Node.js v22.9.0. I'll try upgrading and see if there's a difference.