@nzakas 2025-04-29
When I upgraded Node.js I found that there wasn't much difference at all. :sad: So we can close this.